Finish goal 4,

See merge request ece651-group1/ece651-risk!42

+++ shared/src/test/java/edu/duke/ece651/risk/shared/MapTextViewTest.java
@@ -33,7 +33,7 @@ public class MapTextViewTest {
     idToColor.put((long)1, c2);
     int count = 0;
     for (Territory territory : map.getContinent()) {
-      territory.tryChangeOwnerTo(count % 2);
+      territory.tryChangeOwnerTo((long) (count % 2));
       territory.tryAddUnit(units.get(count));
       count++;
     }
